Clean-up in network.go
- Message functions is divided into two new files, networkSend.go and
neworkResponse.go.
- Added build go command to Dockerfile for CLI. /cli will now run the
  cli-app from the terminal.
- Added "clirun" command to the makefile which takes "id" as argument.
  "id" can be the docker container id or name. Example: "make clirun
  id=196308f5123b" will run the cli-app on container 196308f5123b.
