+++ docker-compose.yml
@@ -13,7 +13,7 @@ services:
     tty: true
     deploy:
       mode: replicated
-      replicas: 25  
+      replicas: 50  
 #        resources:
 #           limits:
 #              cpus: "0.1"
