--- putfile.txt
@@ -1,1 +1,0 @@
-This is a file that we want to store in the kademlia network! :D
